chore/fix: add CI for more platforms and targets, fix build error on 32bit systems

+++ meadowcap/src/mc_capability.rs
@@ -367,7 +367,7 @@ pub(super) mod encoding {
                 }
             }
 
-            let delegations_count = self.delegations_len();
+            let delegations_count = self.delegations_len() as u64;
 
             if delegations_count >= 4294967296 {
                 header |= 0b0011_1111;
@@ -394,7 +394,7 @@ pub(super) mod encoding {
             };
 
             if delegations_count >= 60 {
-                encode_compact_width_be(delegations_count as u64, consumer).await?;
+                encode_compact_width_be(delegations_count, consumer).await?;
             }
 
             let mut prev_area = out.clone();
